Added PostgreSQL Stored Procs

+++ server-postgres/database/migrations/sqls/20221023094536-sqltables-up.sql
@@ -5,8 +5,8 @@ CREATE TABLE IF NOT EXISTS "users" (
     "email" VARCHAR(255) NOT NULL,
     "password" VARCHAR(255) NOT NULL,
     "name" VARCHAR(255) NOT NULL,
-    "createdAt" INT NOT NULL,
-    "updatedAt" INT NOT NULL
+    "createdAt" BIGINT NOT NULL,
+    "updatedAt" BIGINT NOT NULL
 );
 
 -- TABLE FOR TREES
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS "trees" (
     "treeId" SERIAL PRIMARY KEY,
     "name" VARCHAR(255) NOT NULL,
     "description" VARCHAR(255) NOT NULL,
-    "createdAt" INT NOT NULL,
+    "createdAt" BIGINT NOT NULL,
     --   LOCATION POSTGIS POINT
     "location" GEOMETRY,
     "userId" VARCHAR(32) NOT NULL
